test_schutil 相关测试用例完成并通过测试

+++ cpp/schutil_impl.h
@@ -66,8 +66,8 @@ namespace crx
     public:
         void event_callback(uint32_t events);
 
-        std::list<int> m_signals;		//同一个事件可以由多个信号源发送多个不同的信号
-        std::function<void(int)> m_f;
+        std::list<int64_t> m_signals;		//同一个事件可以由多个信号源发送多个不同的信号
+        std::function<void(int64_t)> m_f;
     };
 
     class udp_ins_impl : public eth_event
